将context单独放置目录
