+++ src/app/admin/components/table/EidtableRow.tsx
@@ -5,6 +5,7 @@ import {Input} from '@/components/ui/input';
 import {Label} from '@/components/ui/label';
 import {RadioGroup, RadioGroupItem} from '@/components/ui/radio-group';
 import {Textarea} from '@/components/ui/textarea';
+import {AdminContext} from '@/contexts';
 import {useBoolean} from '@/hooks/useBoolean';
 import {ReportData, UpdateParam} from '@/types';
 import {formatDisplayDateTime} from '@/utils';
@@ -12,7 +13,6 @@ import {zodResolver} from '@hookform/resolvers/zod';
 import {useContext} from 'react';
 import {useForm} from 'react-hook-form';
 import {z} from 'zod';
-import {AdminContext} from '../../page';
 
 interface EidtableRowProps {
     data: ReportData;
