+++ src/app/user/page.tsx
@@ -3,33 +3,24 @@ import Header from '@/components/Header';
 import {useAuth} from '@/hooks/useAuth';
 import {Position} from '@/types';
 import dynamic from 'next/dynamic';
-import {createContext, useState} from 'react';
+import {useState} from 'react';
 import InfoForm from './components/InfoForm';
 import Notice from './components/Notice';
 const MapContainer = dynamic(() => import('./components/MapContainer'), {ssr: false});
-
-export interface PositionContextType {
-    currentPosition: Position;
-    updateCurrentPosition: (newPosition: Position) => void;
-}
-
-export const PositionContext = createContext<PositionContextType>({
-    currentPosition: [116.308303, 39.988792],
-    updateCurrentPosition: () => {},
-});
+import {UserContext} from '@/contexts';
 
 export default function User() {
     useAuth('user');
     const [currentPosition, setCurrentPosition] = useState<Position>([116.308303, 39.988792]);
 
     return (
-        <PositionContext.Provider value={{currentPosition, updateCurrentPosition: setCurrentPosition}}>
+        <UserContext.Provider value={{currentPosition, updateCurrentPosition: setCurrentPosition}}>
             <div>
                 <Header title="违停情况学生反馈提交页面" />
                 <MapContainer />
                 <InfoForm />
                 <Notice />
             </div>
-        </PositionContext.Provider>
+        </UserContext.Provider>
     );
 }
