+++ src/contexts/GuardContext.ts
@@ -1,0 +1,14 @@
+import {Position, UpdateParam} from '@/types';
+import {createContext} from 'react';
+
+export interface GuardContextType {
+    selectedPositions: Position[];
+    updateSelectedPositions: (newPosition: Position[]) => void;
+    updateData: (report: UpdateParam) => Promise<void>;
+}
+
+export const GuardContext = createContext<GuardContextType>({
+    selectedPositions: [],
+    updateSelectedPositions: () => {},
+    updateData: async () => {},
+});
